Fix commit/separate components
