+++ app/src/main/java/com/example/m_banking/presentation/screen/addTransaction/AddTransactionScreen.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.example.m_banking.R
+import com.example.m_banking.presentation.components.ScrollableColumn
 import com.example.m_banking.presentation.theme.ButtonBackground
 import com.example.m_banking.presentation.theme.Typography
 
@@ -166,17 +167,4 @@ fun AddTransactionScreen(navController: NavHostController) {
             }
         }
     }
-}
-
-@Composable
-fun ScrollableColumn(
-    modifier: Modifier = Modifier,
-    content: @Composable ColumnScope.() -> Unit
-) {
-    Column(
-        modifier = modifier
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.spacedBy(16.dp),
-        content = content
-    )
 }
